+++ webClient/src/app/core/menu-bar/menu-bar.config.ts
@@ -238,6 +238,23 @@ export const MENU = [
         }
       }
     ]
+  },
+  {
+    name: 'Session',
+    children: [
+      {
+        name: 'Save Session',
+        action: {
+          internalName: 'saveSession'
+        }
+      },
+      {
+        name: 'Switch Session...',
+        action: {
+          internalName: 'showSessionPicker'
+        }
+      }
+    ]
   }
   // {
   //     name: 'Language Server',